Do not hardcode master (#128)
master does not exist in all repos. For example
https://github.com/ProtonMail/WebClient only has one branch: `public`. Our
GraphQL backend accepts the empty string to mean the default branch. For
constructing the github URL its a bit trickier, instead we use `HEAD` which is
the terminology used by git. For example
https://github.com/ProtonMail/WebClient/blob/HEAD/src/app/elements/index.js is
a valid link to the default branch.

I noticed this issue when browsing errors on gitserver, and noticed `master`
being specified. This was the only repo I could find where this is done outside
of comments.